+++ src/records/PodcastGeneralSettingsRecord.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * @copyright Copyright © vnali
+ */
+
+namespace vnali\studio\records;
+
+use craft\db\ActiveRecord;
+
+/**
+ * Podcast general setting record.
+ *
+ * @property int $podcastId
+ * @property bool $publishRSS
+ */
+class PodcastGeneralSettingsRecord extends ActiveRecord
+{
+    public static function tableName(): string
+    {
+        return '{{%studio_podcast_general_settings}}';
+    }
+}