+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added a podcast general settings page
+- Added an option to disallow publishing RSS for a podcast
 - Fixed a bug where changing lightswitch inputs on the sidebar of the episode edit page, creates unwanted drafts
 
 ## 0.1.0 - 2023-03-02